Depend on DBD::mysql for dpkg and rpm packages
Since innobackupes doesn't depend anymore on the command line client for
mysql, remove the dependency on it for the percona-xtrabackup package,
and move it to the test package which still uses it.

For the percona-xtrabackup package, depend on DBD::mysql which is now the
method innobackup uses (libdbd-mysql-perl in debians).