Bug #1190716: innobackupex  --safe-slave-backup hangs when used on               master
The problem was in uninitialized value usage in wait_for_safe_slave()
when the server is not a replication slave, i.e. when no values in
$con->{slave_status} are defined.

Fixed by moving the check for $con->{slave_status}->{Slave_SQL_Running}
after the check for $host_is_slave and checking if Slave_SQL_Running is
defined when setting $host_is_slave for extra-safety.