Bug #1217426: Empty directory is not backed when stream is used
The problem was that innobackupex ignored empty directories when
creating streaming backups. Normally a database created by the server
contains a ‘db.opt’ file, and so the corresponding directory is not
empty. However, in some circumstances a database may be created with
‘mkdir’. This is what mysql_install_db does to create the default ‘test’
database. Since there are no files to send into the
stream, the resulting stream would contain no references to the empty
directory, so it would not be created when the stream is unpacked and
thus, missing in the backup. Which resulted in inconsistency between the
original data directory and the one after restoring from the backup.

One solution to this problem would be to stream empty directories. This
is rather complicated from the implementation point of view. Since both
xbcrypt and xbstream only deal with files, it would require implementing
support for directory units, i.e. introducing new ‘directory’ chunk
types, bumping the format version, modifications to the datasink API and
utility command line interface, etc.

A much easier solution is to stream a fake empty ‘db.opt’ file for empty
directories. This way the corresponding database will be present in the
backup, and the empty db.opt file will be ignored by the server after
restore.