Implementation of https://blueprints.launchpad.net/percona-xtrabackup/+spec/links-to-percona-site
The idea of the patch is to make build.sh detect the distribution
automatically and patch innobackupex so that the corresponding URL is
printed along with the version text.

This requires:

1. “Configuring” innobackupex, i.e. converting it to a template that
will be used by the build system to replace certain placeholders with
given values and generate the real innobackupex. So innobackupex has
been renamed to innobackupex.pl (with executable bit removed), and the
@XB_DISTRIBUTION@ placeholder has been added. The actual value used to
replace the placeholder is defined by the XB_DISTRIBUTION environment
variable. If not set, then build.sh will auto-detect it based on the
“lsb_release -i” output (i.e. the Distributor ID).

2. XB_DISTRIBUTION to be set to “p” for all PXB packages built by
Percona (regardless of the distribution where they are built). This is
handled in BLD-57.